Add Skill creator and skill wizard
