+++ src/__tests__/components/SkillCreator/SkillWizard.js
@@ -1,16 +1,18 @@
 import React from 'react';
-import SkillCreator from '../../../components/cms/SkillCreator/SkillCreator';
+import SkillWizard from '../../../components/cms/SkillCreator/SkillWizard';
 import { shallow } from 'enzyme';
 import configureMockStore from 'redux-mock-store';
 import { Provider } from 'react-redux';
 const mockStore = configureMockStore();
 const store = mockStore({});
 
-describe('<SkillCreator />', () => {
-  it('render SkillCreator without crashing', () => {
+describe('<SkillWizard />', () => {
+  it('render SkillWizard without crashing', () => {
     shallow(
       <Provider store={store}>
-        <SkillCreator />
+        <SkillWizard
+          location={{ pathname: '/skills/:category/:skill/edit/:lang' }}
+        />
       </Provider>,
     );
   });
